Rename components
TokensPageContainer -> DelegationPage and moved this component to
`/pages/delegation/index` to keep consistency in app structure.
